+++ packages/core-js/actual/async-disposable-stack/index.js
@@ -1,12 +1,8 @@
 'use strict';
-require('../../modules/es.error.cause');
-require('../../modules/es.error.to-string');
-require('../../modules/es.object.to-string');
-require('../../modules/es.promise');
+var parent = require('../../stable/async-disposable-stack');
 require('../../modules/esnext.suppressed-error.constructor');
 require('../../modules/esnext.async-disposable-stack.constructor');
 require('../../modules/esnext.async-iterator.async-dispose');
 require('../../modules/esnext.iterator.dispose');
-var path = require('../../internals/path');
 
-module.exports = path.AsyncDisposableStack;
+module.exports = parent;